# ais-anomaly-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for querying AIS anomaly detection data",
        "title": "AIS Anomaly Detection API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/anomaly-groups": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "anomaly-groups"
                ],
                "summary": "Get anomaly groups with optional filters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "MMSI filter",
                        "name": "mmsi",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Anomaly type filter",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GeoJSONFeatureCollection"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/anomaly-groups/mmsi/{mmsi}": {
            "get": {
                "tags": [
                    "anomaly-groups"
                ],
                "summary": "Get anomaly groups by MMSI",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "MMSI",
                        "name": "mmsi",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GeoJSONFeatureCollection"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/anomaly-groups/{id}": {
            "get": {
                "tags": [
                    "anomaly-groups"
                ],
                "summary": "Get anomaly group by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Anomaly Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AnomalyGroupWithAnomalies"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/base-stations": {
            "get": {
                "description": "Returns all base stations along the Norwegian coast as GeoJSON",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "base-stations"
                ],
                "summary": "Get all base stations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GeoJSONFeatureCollection"
                        }
                    }
                }
            }
        },
        "/base-stations/{id}": {
            "get": {
                "description": "Returns a single base station by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "base-stations"
                ],
                "summary": "Get base station by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Base Station ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BaseStation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Anomaly": {
            "type": "object",
            "properties": {
                "anomalyGroupId": {
                    "type": "integer",
                    "example": 1
                },
                "createdAt": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "dataSource": {
                    "type": "string",
                    "example": "SYNTHETIC"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "metadata": {
                    "type": "object"
                },
                "signalStrength": {
                    "type": "number",
                    "example": -95.5
                },
                "sourceId": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.AnomalyGroupWithAnomalies": {
            "type": "object",
            "properties": {
                "anomalies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Anomaly"
                    }
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "lastActivityAt": {
                    "type": "string",
                    "example": "2024-01-15T14:30:00Z"
                },
                "latitude": {
                    "type": "number",
                    "example": 59.9139
                },
                "longitude": {
                    "type": "number",
                    "example": 10.7522
                },
                "mmsi": {
                    "type": "integer",
                    "example": 123456789
                },
                "startedAt": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "type": {
                    "type": "string",
                    "example": "speed_anomaly"
                }
            }
        },
        "models.BaseStation": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "latitude": {
                    "type": "number",
                    "example": 58.96463228151012
                },
                "longitude": {
                    "type": "number",
                    "example": 5.738597994470069
                },
                "name": {
                    "type": "string",
                    "example": "Stavanger"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid_request"
                },
                "message": {
                    "type": "string",
                    "example": "The request parameters are invalid"
                }
            }
        },
        "models.GeoJSONFeature": {
            "type": "object",
            "properties": {
                "geometry": {
                    "$ref": "#/definitions/models.GeoJSONGeometry"
                },
                "properties": {
                    "type": "object"
                },
                "type": {
                    "type": "string",
                    "example": "Feature"
                }
            }
        },
        "models.GeoJSONFeatureCollection": {
            "type": "object",
            "properties": {
                "features": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GeoJSONFeature"
                    }
                },
                "type": {
                    "type": "string",
                    "example": "FeatureCollection"
                }
            }
        },
        "models.GeoJSONGeometry": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    },
                    "example": [
                        10.7522,
                        59.9139
                    ]
                },
                "type": {
                    "type": "string",
                    "example": "Point"
                }
            }
        }
    }
}